Rename files: trackingDog to TrackingDog
