+++ lib/TrackingDog.js
