+++ lib/TrackingDog.spec.js
@@ -1,5 +1,5 @@
 const expect = require('unexpected');
-const TrackingDog = require('./trackingDog');
+const TrackingDog = require('./TrackingDog');
 
 describe('trackingDog', function() {
   describe('with a local file path', function() {
