+++ lib/cli.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const TrackingDog = require('./trackingDog');
+const TrackingDog = require('./TrackingDog');
 const MagicPen = require('magicpen');
 const urlTools = require('urltools');
 const pathModule = require('path');
